moved OGR to filters/hexbin
